:recycle: to structural project
